spark-apps/requirements.txt
kafka-python==2.0.2
redis==4.5.4
numpy==1.24.3